@@ -2,6 +2,7 @@
     "name": "vite-test",
     "private": true,
     "version": "0.0.0",
+    "type": "module",
     "scripts": {
         "dev": "vite",
         "build": "vite build",
@@ -26,7 +27,7 @@
         "vite": "^5.4.2",
         "vite-plugin-svgr": "^4.3.0",
         "vite-svg-loader": "^5.1.0",
-        "vue": "3.4",
+        "vue": "^3.5.13",
         "vue-slider-component": "^4.1.0-beta.7"
     }
 }
